pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "sunschool"
version = "0.1.0"
description = "AI tutoring platform for kids"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110,<1",
    "uvicorn[standard]>=0.29,<1",
    "psycopg2-binary>=2.9,<3",
    "httpx>=0.27,<1",
    "google-auth>=2.29,<3",
    "pydantic>=2.7,<3",
    "pydantic-settings>=2.2,<3",
    "python-dotenv>=1.0,<2",
    "networkx>=3.3,<4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1,<9",
    "pytest-asyncio>=0.23,<1",
    "ruff>=0.4,<1",
    "mypy>=1.10,<2",
]

# ---------------------------------------------------------------------------
# Tools
# ---------------------------------------------------------------------------

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
